Move functions over to citation module.
